Fixed critical issue where upgrading the db causes getWritableDatabase() to be called recursively. For future reference, no locking on the db object is needed, and the db handle given to onUpgrade() should be used for any db-related upgrade operations.
Explanation: getWritableDatabase() will invoke onUpgrade() with the db handle it just acquired if the database needs to be upgraded. Attempting to acquire the db handle again lead to a guaranteed deadlock with the old locking code, and to a recursive calll to getWritableDatabase() with the code in the prior commit.