Reenabled downloading the images for campaigns in FeedbackService, except that it now downloads them to a folder called "aw_images_cached", with the same structure as the original "aw_images" folder.
Note that the cache is never purged, and if a photo already exists with the same UUID it will never be replaced. Also note that the service assumes the files are pngs -- I have to figure out a way to know the type of the file prior to the request or else it's hard to know what filename to check for. Solution: just redownload the file anyway, or possibly check after downloading the file and discard the downloaded data if the photo already exists.