Fixed #89 logout activity flow bugs
The LoginActivity is now the starting point for the app. It decides if it should show the login screen
or go to the dashboard. For the task clear to work, the LoginActivity continues to run until the
dashboard finishes at which point it finishes as well. The way the LoginActivity waits for the
DashboardActivity prevents the Dashboard from having the singleInstanceState flag (otherwise the
onActivityResult callback doesn't work). This means anytime we want the user to go back to the
dashboard we have to pass the FLAG_ACTIVITY_CLEAR_TOP (as I changed in the ActionBar). The
LauncherActivity is not used any more. We will have to reintroduce it (or something like it) so we can
enable single campaign mode.