moved the campaign download to the dashboard and added a primitive timestamp check to make sure we don't try to download back to back
We can probably make this a lot better but it works for now. It has a simple check of the time when the
dashboard starts (which is stored in the shared preferences). If it has been less than 5 minutes since the
last time we tried to download campaigns we will download them again (it doesn't matter if the call succeeded
or not, it wont try again for 5 minutes). We could also change the async task. Right now the same task could
run multiple times. (if the user clicks the refresh button a lot for example). This can cause bad behavior
in the db.