Closes issue #233 ("going from survey info to reminders doesn't auto-select the currently-viewed survey"), albeit somewhat clumsily.
To elaborate: on the survey info page, clicking the reminders button sends the currently viewed survey's name as an intent to the trigger list activity. The trigger list activity then passes it along to any trigger creation activities that are opened from the trigger list, which requires that each derived trigger type must handle the extra containing the default action selections.

In the future, I'd like to replace the trigger list activity with a more standard campaign/survey filtered list. The default survey selection for new reminders would then be the survey on which you're currently filtering, if one is selected, or nothing if "All Surveys" is selected.