Surveys now show up in the survey list in the order that they're defined in the xml. Closes #231.
Note that I'm sorting on the autoincrementing id for the surveys, which works because the surveys are inserted in the order that they're parsed from the xml. If this were to someday change (for instance, if the surveys were removed and then reinserted without parsing them from the xml), this might break. I don't anticipate this happening, just more of an fyi.