fixed #218. Custom choices for campaigns are deleted when a campaign is removed.
The avatar icon for the campaign as well as any images which were downloaded are also
deleted. The only thing that isn't deleted when a campaign is removed is triggers
since people have expressed an interest that they stay so they don't have to put them
in everytime. It seems ok for them to stay since they are independent of campaigns