fixed #263 Updated ohmage project to use the account manager
The account manager now keeps track of the auth token returned by the server. The
AccountHelper class contains the methods which deals with the current account.
The account will show up in the accounts settings page in the android settings,
but the user can't remove it. We need to see if there is a way to prevent it from
showing up there.